// src/app/core/models/modelsTest/movieResponseApi.ts
export interface MovieResponse {
  adult: boolean;
  backdrop_path: string;
  id: number;
  title: string;
  original_language: string;
  original_title: string;
  overview: string;
  poster_path: string;
  media_type: string;
  genre_ids: number[];
  popularity: number;
  release_date: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface MovieResponseApi {
  page: number;
  results: MovieResponse[];
  total_pages: number;
  total_results: number;
}